+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+- Handle whitespace in HTML-strings using ERB-tags
+
 ## [0.9.1] - 2023-06-28
 
 - Handle formatting of multi-line ERB-tags with more than one statement.